Create almanax.py
It's very easy to run a bot on your server, so I won't be doing that part. All the code is here and simply start the program, either on your computer or a virtual machine. I didn't try it yet, but I guess you won't have to log into your discord account as it's running with the bot token, and not with a logged-in discord account.